+++ src/interfaces/IMetaMorpho.sol
@@ -27,7 +27,7 @@ interface IMetaMorpho is IERC4626 {
 
     function fee() external view returns (uint96);
     function feeRecipient() external view returns (address);
-    function timelock() external view returns (uint256);
+    function timelock() external view returns (uint96);
     function supplyQueue(uint256) external view returns (Id);
     function withdrawQueue(uint256) external view returns (Id);
     function config(Id) external view returns (uint192 cap, uint64 withdrawRank);
